[BugFix][compiler-rt] Fix OHOS HWASan and GWP-ASan build issues
Adaptations:
- Local 21.x fix; no OHOS 15.0.4 source commit.
- hwasan: exclude OHOS from detect_leaks default (CAN_SANITIZE_LEAKS is 0).
- hwasan_allocator: treat OHOS like Android for AP64 kSpaceSize selection.
- gwp_asan: cast PersistInterval/ReserveCounter average to unsigned for -Wshorten-64-to-32.

+++ compiler-rt/lib/gwp_asan/guarded_pool_allocator.h
@@ -109,8 +109,10 @@ public:
 #if defined(OHOS_LLVM) && defined(__OHOS__)
     Nmalloc++;
     if (Nmalloc % PRINT_COUNTER == 0) {
-      unsigned AvgDurationUs =
-          ReserveCounter ? PersistInterval / ReserveCounter : 0;
+      // PersistInterval/ReserveCounter are size_t; explicit cast for
+      // -Werror,-Wshorten-64-to-32 on 64-bit OHOS.
+      unsigned AvgDurationUs = static_cast<unsigned>(
+          ReserveCounter ? PersistInterval / ReserveCounter : 0);
       MUSL_LOG("[gwp_asan]: AvgDuration %{public}u us, FreeSlotsLength "
                "%{public}d\n",
                AvgDurationUs, FreeSlotsLength);
